Initial commit of dgoss. Add --color flag. Add more template functions (not documented yet)
